feat: entity id 타입 변경

+++ src/main/java/com/wanted/onboarding/domain/recruitmentnotice/repository/RecruitmentNoticeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecruitmentNoticeRepository extends JpaRepository<RecruitmentNotice, Integer> {
+public interface RecruitmentNoticeRepository extends JpaRepository<RecruitmentNotice, Long> {
 
 }
